gtp: add socket destroy handler
Add a socket destroy handler and use it to detach and release all
resources from an UDP socket.

Do not longer pin the socket with a reference. This together with
the release handler allows user space to close the socket from
underneath us. The destroy handler will handle the rest.