gtp: select netns based on NL attribute
This permits a split namespace setup where the GTP transport
sockets are in one namespace the gtp tunnel interface is in
another namespace.

The target namespece is selected by the new GTPA_NET_NS_FD NL
attributes. It fall back to the netns of the GTP-U sockets if
the NL attr is not present.